Correction to first comments of Suprevisor for the module5-task-1 (removed unused modules, rename from library to data module, rename the folder for modules to modules)
