+++ js/modules/get-random.js
